feat: improve player movement

+++ src/engine/Renderable.cpp
@@ -8,7 +8,7 @@ Renderable::Renderable()
     : x_(0), y_(0), w_(0), h_(0), scaleX_(1.f), scaleY_(1.f), frames_(1), sides_(1),
       currentFrame_(0), texture_(NULL), spriteSide_(lookAt::left) {}
 
-Renderable::Renderable(int x, int y)
+Renderable::Renderable(float x, float y)
     : x_(x), y_(y), w_(0), h_(0), scaleX_(1.f), scaleY_(1.f), frames_(1), sides_(1),
       currentFrame_(0), texture_(NULL), spriteSide_(lookAt::left) {}
 
@@ -69,7 +69,7 @@ void Renderable::render(int x, int y, SDL_Renderer* renderer) {
     SDL_RenderCopy(renderer, this->texture_, &srcRenderQuad, &dstRenderQuad);
 }
 
-void Renderable::move(int x, int y) {
+void Renderable::move(float x, float y) {
     this->x_ += x;
     this->y_ += y;
 }
